Move BulkDescribeUtils unit to utils app tests
https://github.com/MTG/freesound/pull/1093
