+++ accounts/tests.py
@@ -1283,226 +1283,3 @@ class AboutFieldVisibilityTests(object):  # temporarily disable this test becaus
         self.client.login(username='admin', password='testpass')
         self._check_visibility('spammer', True)
         self._check_visible()
-
-
-class BulkDescribe(TestCase):
-
-    fixtures = ['initial_data']
-
-    @staticmethod
-    def create_audio_files(filenames, base_path):
-        for filename in filenames:
-            f = open(base_path + filename, 'a')
-            f.write(os.urandom(1024))  # Add random content to the file to avoid equal md5
-            f.close()
-
-    @staticmethod
-    def create_csv_file(filename, lines, base_path):
-        csv_file_path = '%s/%s' % (base_path, filename)
-        csv_fid = open(csv_file_path, 'w')
-        for line in lines:
-            csv_fid.write(line + '\n')
-        csv_fid.close()
-        return csv_file_path
-
-    @override_settings(UPLOADS_PATH=tempfile.mkdtemp())
-    @override_settings(CSV_PATH=tempfile.mkdtemp())
-    def test_validate_input_csv_file(self):
-        # Create user uploads folder and test audio files
-        user = User.objects.create_user("testuser", password="testpass")
-        user_upload_path = settings.UPLOADS_PATH + '/%i/' % user.id
-        os.mkdir(user_upload_path)
-        self.create_audio_files(['file1.wav', 'file2.wav', 'file3.wav', 'file4.wav', 'file5.wav'], user_upload_path)
-
-        # Create CSV files folder with descriptions
-        csv_file_base_path = settings.CSV_PATH + '/%i/' % user.id
-        os.mkdir(csv_file_base_path)
-
-        # Test CSV with al lines and metadata ok
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name',
-            'file1.wav;New name for file1.wav;tag1 tag2 tag3;41.4065, 2.19504, 23;'
-            '"Description for file";Creative Commons 0;ambient',  # All fields valid
-            'file2.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;',  # Only mandatory fields
-            'file3.wav;;tag1 tag2 tag3;;'
-            '"Description for file";Creative Commons 0;ambient',  # All mandatory fields and some optional fields
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path, username=user.username)
-        self.assertEqual(len(global_errors), 0)  # No global errors
-        self.assertEqual(len([line for line in lines_validated if line['line_errors']]), 0)  # No line errors
-
-        # Test username does not exist
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path, username="unexisting username")
-        self.assertEqual(len(global_errors), 0)  # No global errors
-        self.assertEqual(len([line for line in lines_validated if line['line_errors']]), 3)  # Three line errors
-        self.assertTrue('username' in lines_validated[0]['line_errors'])  # User does not exist error reported
-        self.assertTrue('username' in lines_validated[1]['line_errors'])  # User does not exist error reported
-        self.assertTrue('username' in lines_validated[2]['line_errors'])  # User does not exist error reported
-
-        # Test missing/duplicated audiofile and wrong number of rows
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name',
-            'file1.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;',  # File exists, fields ok
-            'file2.wav;;tag1 tag2 tag3;;;Creative Commons 0;',  # Missing description
-            'file3.wav;;tag1 tag2 tag3;;"Description for file";',  # Wrong number of columns
-            'file6.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;',  # Audiofile does not exist
-            'file2.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;',  # Audiofile already described
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path, username=user.username)
-        self.assertEqual(len(global_errors), 0)  # No global errors
-        self.assertEqual(len([line for line in lines_validated if line['line_errors']]), 4)  # Four lines have errors
-        self.assertTrue('description' in lines_validated[1]['line_errors'])  # Missing description error reported
-        self.assertTrue('columns' in lines_validated[2]['line_errors'])  # Wrong number of columns reported
-        self.assertTrue('audio_filename' in lines_validated[3]['line_errors'])  # Audiofile not exist error reported
-        self.assertTrue('audio_filename' in lines_validated[4]['line_errors'])  # File already described error reported
-
-        # Test validation errors in individual fields
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name',
-            'file1.wav;;tag1 tag2;;"Description for file";Creative Commons 0;',  # Wrong tags (less than 3)
-            'file2.wav;;tag1,tag2;;"Description for file";Creative Commons 0;',  # Wrong tags (less than 3)
-            'file3.wav;;tag1,tag2;gr87g;"Description for file2";Creative Commons 0;',  # Wrong geotag
-            'file4.wav;;tag1,tag2;42.34,190.45,15;"Description for file";Creative Commons 0;',  # Wrong geotag
-            'file5.wav;;tag1 tag2 tag3;;"Description for file";Sampling+;',  # Invalid license
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path, username=user.username)
-        self.assertEqual(len(global_errors), 0)  # No global errors
-        self.assertEqual(len([line for line in lines_validated if line['line_errors']]), 5)  # Five lines have errors
-        self.assertTrue('tags' in lines_validated[0]['line_errors'])  # Wrong tags
-        self.assertTrue('tags' in lines_validated[1]['line_errors'])  # Wrong tags
-        self.assertTrue('geotag' in lines_validated[2]['line_errors'])  # Wrong geotag
-        self.assertTrue('geotag' in lines_validated[3]['line_errors'])  # Wrong geotag
-        self.assertTrue('license' in lines_validated[4]['line_errors'])  # Wrong license
-
-        # Test wrong header global errors
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;unknown_field',
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path, username=user.username)
-        self.assertEqual(len(global_errors), 1)  # One global error
-        self.assertTrue('Invalid header' in global_errors[0])  # Invalid header error reported
-
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name',
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path,
-                                    username=None)  # Not passing username, header should now include 'username' field
-        self.assertEqual(len(global_errors), 1)  # One global error
-        self.assertTrue('Invalid header' in global_errors[0])  # Invalid header error reported
-
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name;username',
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = \
-            validate_input_csv_file(header, lines, user_upload_path,
-                                    username=None)  # Not passing username, header should now include 'username' field
-        self.assertEqual(len(global_errors), 0)  # No global errors
-
-        # Test username errors when not passing username argument to validate_input_csv_file
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name;username',
-            'file1.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;;new_username',  # User does not exist
-            'file2.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;',  # Invlaid num columns
-            'file3.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;;testuser',  # All fields OK
-        ], csv_file_base_path)
-        header, lines = get_csv_lines(csv_file_path)
-        lines_validated, global_errors = validate_input_csv_file(header, lines, user_upload_path, username=None)
-        self.assertEqual(len(global_errors), 0)  # No global errors
-        self.assertEqual(len([line for line in lines_validated if line['line_errors']]), 2)  # Two lines have errors
-        self.assertTrue('username' in lines_validated[0]['line_errors'])  # User does not exist
-        self.assertTrue('columns' in lines_validated[1]['line_errors'])  # Invalid number of columns
-
-        # Delete tmp directories
-        shutil.rmtree(settings.UPLOADS_PATH)
-        shutil.rmtree(settings.CSV_PATH)
-
-    @override_settings(UPLOADS_PATH=tempfile.mkdtemp())
-    @override_settings(CSV_PATH=tempfile.mkdtemp())
-    def test_bulk_describe_from_csv(self):
-
-        # Create user uploads folder and test audio files
-        user = User.objects.create_user("testuser", password="testpass")
-        user_upload_path = settings.UPLOADS_PATH + '/%i/' % user.id
-        os.mkdir(user_upload_path)
-        self.create_audio_files(['file1.wav', 'file2.wav', 'file3.wav', 'file4.wav', 'file5.wav'], user_upload_path)
-
-        # Create CSV files folder with descriptions
-        csv_file_base_path = settings.CSV_PATH + '/%i/' % user.id
-        os.mkdir(csv_file_base_path)
-
-        # Create Test CSV with some lines ok and some wrong lines
-        csv_file_path = self.create_csv_file('test_descriptions.csv', [
-            'audio_filename;name;tags;geotag;description;license;pack_name',
-            'file1.wav;;tag1 tag2 tag3;41.4065, 2.19504, 23;"Description for file";Creative Commons 0;ambient',  # OK
-            'file2.wav;;tag1 tag2 tag3;;"Description for file";Invalid license;',  # Invalid license
-            'file3.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0',  # Wrong number of columns
-            'file4.wav;;tag1 tag2 tag3;dg;"Description for file";Creative Commons 0;',  # Invalid geotag
-            'file5.wav;;tag1 tag2 tag3;;"Description for file";Creative Commons 0;',  # OK
-        ], csv_file_base_path)
-
-        # Test case when no sounds are been created because CSV file has some errors and 'force_import' is set to False
-        bulk_describe_from_csv(csv_file_path,
-                               delete_already_existing=False,
-                               force_import=False,
-                               sounds_base_dir=user_upload_path,
-                               username=user.username)
-        self.assertEqual(user.sounds.count(), 0)  # User has no sounds
-
-        # Test case using 'force_import' (only sounds for lines that validate ok will be created)
-        bulk_describe_from_csv(csv_file_path,
-                               delete_already_existing=False,
-                               force_import=True,
-                               sounds_base_dir=user_upload_path,
-                               username=user.username)
-        self.assertEqual(user.sounds.count(), 2)  # The two sounds that had correct metadata have been added
-        sound1 = Sound.objects.get(user=user, original_filename='file1.wav')  # Get first correct sound
-        sound1_id = sound1.id  # This is used in a test below
-        self.assertTrue(sound1.geotag)  # Check sound has geotag object assigned
-        self.assertEquals(sound1.pack.name, 'ambient')  # Check sound has pack and name of pack is 'ambient'
-        sound2 = Sound.objects.get(user=user, original_filename='file5.wav')  # Get last correct sound
-        sound2_id = sound2.id  # This is used in a test below
-        self.assertIsNone(sound2.geotag)  # Check sound has no geotag
-        self.assertIsNone(sound2.pack)  # Check sound has no pack
-
-        # Run again using 'force_import' and sounds won't be created because sounds already exist and md5 check fails
-        # NOTE: first we copy back the files that were already successfully added because otherwise these don't exist
-        shutil.copy(sound1.locations()['path'], os.path.join(user_upload_path, 'file1.wav'))
-        shutil.copy(sound2.locations()['path'], os.path.join(user_upload_path, 'file5.wav'))
-        bulk_describe_from_csv(csv_file_path,
-                               delete_already_existing=False,
-                               force_import=True,
-                               sounds_base_dir=user_upload_path,
-                               username=user.username)
-        self.assertEqual(user.sounds.count(), 2)  # User still has two sounds, no new sounds added
-
-        # Run again using 'force_import' AND 'delete_already_existing' and existing sounds will be removed before
-        # creating the new ones
-        # NOTE: first we copy back the files that failed MD5 check as files are discarted (deleted) when MD5 fails
-        shutil.copy(sound1.locations()['path'], os.path.join(user_upload_path, 'file1.wav'))
-        shutil.copy(sound2.locations()['path'], os.path.join(user_upload_path, 'file5.wav'))
-        bulk_describe_from_csv(csv_file_path,
-                               delete_already_existing=True,
-                               force_import=True,
-                               sounds_base_dir=user_upload_path,
-                               username=user.username)
-        self.assertEqual(user.sounds.count(), 2)  # User still has two sounds
-        new_sound1 = Sound.objects.get(user=user, original_filename='file1.wav')  # New version of first correct sound
-        new_sound2 = Sound.objects.get(user=user, original_filename='file5.wav')  # New version of last correct sound
-        self.assertNotEqual(new_sound1.id, sound1_id)  # Check that IDs are not the same
-        self.assertNotEqual(new_sound2.id, sound2_id)  # Check that IDs are not the same
-
-        # Delete tmp directories
-        shutil.rmtree(settings.UPLOADS_PATH)
-        shutil.rmtree(settings.CSV_PATH)
